refactor: extract socket setup logic into useSetupSocketEvents hook
